Dirty

+++ industrial_sem2/compilier/ast.hpp
@@ -54,7 +54,7 @@ struct ast_node_t
 
     union
     {
-        int number;
+        int64_t number;
         string_view_t ident;
     };
